@@ -13,27 +13,26 @@
   "private": true,
   "devDependencies": {
     "@angular-devkit/architect": "^0.11.2",
-    "@angular-devkit/build-angular": "^0.13.8",
+    "@angular-devkit/build-angular": "~0.803.29",
     "@angular-devkit/build-optimizer": "^0.11.2",
     "@angular-devkit/core": "^7.1.2",
-    "@angular/animations": "7.2.4",
+    "@angular/animations": "8.2.14",
     "@angular/cdk": "7.1.1",
-    "@angular/cli": "7.3.1",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/core": "7.2.4",
-    "@angular/elements": "7.2.4",
+    "@angular/cli": "8.3.29",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/elements": "8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "7.2.4",
-    "@angular/http": "7.2.4",
-    "@angular/language-service": "7.2.4",
+    "@angular/forms": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@angular/material": "7.1.1",
     "@angular/material-moment-adapter": "^7.1.1",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
-    "@angular/service-worker": "7.2.4",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@angular/service-worker": "8.2.14",
     "@ngtools/webpack": "^7.3.1",
     "@ngxs/logger-plugin": "^3.4.3",
     "@ngxs/store": "^3.4.3",
@@ -44,7 +43,7 @@
     "@types/node": "^10.12.12",
     "angular2gridster": "^7.0.2",
     "chart.js": "^2.7.2",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "concat": "^1.0.3",
     "core-js": "^2.6.0",
     "css-element-queries": "1.1.1",
@@ -81,7 +80,7 @@
     "tslint": "^5.11.0",
     "typeface-ubuntu": "0.0.65",
     "typeface-ubuntu-mono": "0.0.54",
-    "typescript": "~3.2.4",
-    "zone.js": "0.8.29"
+    "typescript": "~3.5.3",
+    "zone.js": "~0.9.1"
   }
 }
