@@ -39,7 +39,7 @@ import { NavigatorPanelComponent } from '../navigator-panel/navigator-panel.comp
 export class SettingsPanelComponent implements OnInit {
     //@HostBinding('class.settings-navigator') private _hostClass = true;
 
-    @ViewChild(NavigatorPanelComponent) private navPanel: NavigatorPanelComponent;
+    @ViewChild(NavigatorPanelComponent, { static: false }) private navPanel: NavigatorPanelComponent;
 
     @Input() activeNavSection: any = '';
     @Input() drawerMode: any = 'over';
