@@ -128,9 +128,9 @@ export class DashboardComponent implements OnInit, OnDestroy {
     @Select(NavigatorState.getDrawerOpen) drawerOpen$: Observable<any>;
     @Select(EventsState.GetEvents) events$: Observable<any>;
 
-    @ViewChild('tplVariablePanel') tplVariablePanel : TemplateVariablePanelComponent;
+    @ViewChild('tplVariablePanel', { static: false }) tplVariablePanel : TemplateVariablePanelComponent;
     // available widgets menu trigger
-    @ViewChild('availableWidgetsMenuTrigger', { read: MatMenuTrigger }) availableWidgetsMenuTrigger: MatMenuTrigger;
+    @ViewChild('availableWidgetsMenuTrigger', { read: MatMenuTrigger, static: false }) availableWidgetsMenuTrigger: MatMenuTrigger;
 
     get availableWidgetsMenuIsOpen(): boolean {
         if (this.availableWidgetsMenuTrigger) {
@@ -140,11 +140,11 @@ export class DashboardComponent implements OnInit, OnDestroy {
     }
 
     // portal templates
-    @ViewChild('dashboardNavbarTmpl') dashboardNavbarTmpl: TemplateRef<any>;
+    @ViewChild('dashboardNavbarTmpl', { static: true }) dashboardNavbarTmpl: TemplateRef<any>;
 
-    @ViewChild(DboardContentComponent, {read: DboardContentComponent}) dbContent: DboardContentComponent;
+    @ViewChild(DboardContentComponent, { read: DboardContentComponent, static: true }) dbContent: DboardContentComponent;
 
-    @ViewChild(NavbarClipboardMenuComponent, {read: NavbarClipboardMenuComponent}) clipboardMenu: NavbarClipboardMenuComponent;
+    @ViewChild(NavbarClipboardMenuComponent, { read: NavbarClipboardMenuComponent, static: false }) clipboardMenu: NavbarClipboardMenuComponent;
 
     // portal placeholders
     dashboardNavbarPortal: TemplatePortal;
