@@ -53,11 +53,11 @@ export class HeatmapWidgetComponent implements OnInit, AfterViewInit, OnDestroy 
   @Input() mode = 'view'; // view/explore/edit
   @Input() readonly = true;
 
-  @ViewChild('widgetOutputContainer') private widgetOutputContainer: ElementRef;
-  @ViewChild('widgetTitle') private widgetTitle: ElementRef;
-  @ViewChild('widgetoutput') private widgetOutputElement: ElementRef;
-  @ViewChild('graphLegend') private dygraphLegend: ElementRef;
-  @ViewChild('dygraph') private dygraph: ElementRef;
+  @ViewChild('widgetOutputContainer', { static: true }) private widgetOutputContainer: ElementRef;
+  @ViewChild('widgetTitle', { static: false }) private widgetTitle: ElementRef;
+  @ViewChild('widgetoutput', { static: true }) private widgetOutputElement: ElementRef;
+  @ViewChild('graphLegend', { static: true }) private dygraphLegend: ElementRef;
+  @ViewChild('dygraph', { static: true }) private dygraph: ElementRef;
 
   Object = Object;
   private listenSub: Subscription;
