@@ -43,8 +43,8 @@ export class TopnWidgetComponent implements OnInit, OnDestroy, AfterViewInit {
     @Input() mode = 'view'; // view/explore/edit
     @Input() readonly = true;
 
-    @ViewChild('widgetoutput') private widgetOutputElement: ElementRef;
-    @ViewChild('container') private container: ElementRef;
+    @ViewChild('widgetoutput', { static: true }) private widgetOutputElement: ElementRef;
+    @ViewChild('container', { static: true }) private container: ElementRef;
 
     Object = Object;
     private listenSub: Subscription;
