@@ -19,15 +19,15 @@ import { Routes, RouterModule } from '@angular/router';
 import { AuthGuardService } from './auth-guard.service';
 
 const routes: Routes =  [
-  { path: 'd', loadChildren: 'app/dashboard/dashboard.module#DashboardModule' },
-  { path: 'snap', loadChildren: 'app/dashboard/dashboard.module#DashboardModule', canLoad: [ AuthGuardService ] },
-  { path: 'main', loadChildren: 'app/landing-page/landing-page.module#LandingPageModule', canLoad: [ AuthGuardService ]  },
-  { path: 'a', loadChildren: 'app/alerts/alerts.module#AlertsModule', canLoad: [ AuthGuardService ]  },
-  { path: 'user', loadChildren: 'app/user/user.module#UserModule', canLoad: [ AuthGuardService ]  },
-  { path: 'namespace', loadChildren: 'app/namespace/namespace.module#NamespaceModule', canLoad: [ AuthGuardService ]  },
-  { path: 'admin', loadChildren: 'app/admin/admin.module#AdminModule', canLoad: [ AuthGuardService ] },
+  { path: 'd', loadChildren: () => import('app/dashboard/dashboard.module').then(m => m.DashboardModule) },
+  { path: 'snap', loadChildren: () => import('app/dashboard/dashboard.module').then(m => m.DashboardModule), canLoad: [ AuthGuardService ] },
+  { path: 'main', loadChildren: () => import('app/landing-page/landing-page.module').then(m => m.LandingPageModule), canLoad: [ AuthGuardService ]  },
+  { path: 'a', loadChildren: () => import('app/alerts/alerts.module').then(m => m.AlertsModule), canLoad: [ AuthGuardService ]  },
+  { path: 'user', loadChildren: () => import('app/user/user.module').then(m => m.UserModule), canLoad: [ AuthGuardService ]  },
+  { path: 'namespace', loadChildren: () => import('app/namespace/namespace.module').then(m => m.NamespaceModule), canLoad: [ AuthGuardService ]  },
+  { path: 'admin', loadChildren: () => import('app/admin/admin.module').then(m => m.AdminModule), canLoad: [ AuthGuardService ] },
   { path: '', redirectTo: 'main', pathMatch: 'full', canLoad: [ AuthGuardService ]  },
-  { path: 'error', loadChildren:'app/error/error.module#ErrorModule'},
+  { path: 'error', loadChildren:() => import('app/error/error.module').then(m => m.ErrorModule)},
   { path: '**', redirectTo: 'main', pathMatch: 'full', canLoad: [ AuthGuardService ] }
 ];
 
